--- widgets/news/components/index.ts
@@ -1,1 +1,0 @@
-export { NewsList } from './news-list';
