--- widgets/news/components/news-list/index.ts
@@ -1,1 +1,0 @@
-export { NewsList } from './news-list';
